pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "oakutils"
version = "1.5.1"
authors = [
  {name="Justin Davis", email="davisjustin302@gmail.com"},
]
maintainers = [
  {name="Justin Davis", email="davisjustin302@gmail.com"},
]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Natural Language :: English",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Hardware",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: End Users/Desktop",
    "Typing :: Typed",
]
requires-python=">=3.8, <3.13"
dependencies = [
    "depthai>=2.24",
    "numpy>=1.20,<2.0",
    "opencv-contrib-python>=4.9.0",
    "cv2ext>=0.0.10",
    "typing_extensions>=4.9.0",
    "setuptools>=57.0.0",
]

[project.urls]
"Homepage" = "https://github.com/justincdavis/oakutils"
"Bug Tracker" = "https://github.com/justincdavis/oakutils/issues"

[project.optional-dependencies]
o3d = [
    "open3d>=0.18.0",
]
compiler = [
    "kornia>=0.6.0",
    "torch>=1.9.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.15.0",
    "onnxsim>=0.4.33",
    "blobconverter==1.4.3",
    "requests>=2.25.0",
    "boto3>=1.34.37",  # Removes the need for search during install
    "botocore>=1.34.37",  # Removes the need for search during install
]
ci = [
    "pyupgrade>=3.10",
    "black>=24.0",
    "isort>=5.10",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "types-setuptools>=57.0.0",
    "types-requests>=2.25.0",
]
test = [
    "pytest>=6.2.0",
]
docs = [
    "sphinx>=6.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst_parser>=1.0.0",
]
dev = [
    "oakutils[o3d]",
    "oakutils[compiler]",
    "oakutils[ci]",
    "oakutils[test]",
    "oakutils[docs]",
    "twine>=4.0.0",
    "wheel>=0.37.0",
    "bumpver>=2023.1126",
    "pyclean>=2.7.0",
    "pyright>=1.1.348",
]

[tool.bumpver]
current_version = "1.5.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = true
tag             = true
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/oakutils/__init__.py" = ["{version}"]
"docs/conf.py" = ["{version}"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
oakutils = ["src/*.blob", "src/oakutils/py.typed"]

[tool.ruff]
fix = true
show-fixes = true
# Same as Black.
line-length = 88
# Assume Python 3.8
target-version = "py38"
# Other options
builtins = ["_"]

[tool.ruff.lint]
select = ["ALL", "PLR2004"]
ignore = ["PLR", "C901", "E501", "D203", "D212", "ERA"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.mypy]
python_version = "3.8"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false
strict_equality = true
check_untyped_defs = true
disallow_subclassing_any = false
disallow_untyped_decorators = true
disallow_any_generics = false
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_reexport = true
warn_return_any = true

# ignore cv2, open3d, blobconverter
[[tool.mypy.overrides]]
module = [  
    'cv2',
    'open3d',
    'onnxsim',
    'blobconverter',
]
ignore_errors = true

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules",
    "**/__pycache__",
]

ignore = ["scripts/*"]

defineConstant = {DEBUG = true}
stubPath = "typings"
reportMissingImports = true
reportMissingTypeStubs = false
  
pythonVersion = "3.8"
pythonPlatform = "Linux"

executionEnvironments = [
    { root = "src", pythonVersion = "3.8", stubPath = "typings" },
]
